requirements.txt
fastapi
uvicorn
jinja2
reportlab
python-dotenv
pandas
openpyxl
geopy
requests
pycountry
pypdf
